Fix windows compilation error

+++ ddsenabler_participants/src/cpp/RpcUtils.cpp
@@ -204,11 +204,11 @@ UUID generate_UUID()
     UUID uuid;
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_int_distribution<uint8_t> dis(0, 255);
+    std::uniform_int_distribution<unsigned int> dis(0, 255);
 
     for (size_t i = 0; i < sizeof(uuid); ++i)
     {
-        uuid[i] = dis(gen);
+        uuid[i] = static_cast<uint8_t>(dis(gen));
     }
     return uuid;
 }
